se crea la clase utilitaria
