+++ src/main/java/ViewDatosBasicos.java
@@ -97,7 +97,8 @@ class PanelMenu extends JPanel implements ActionListener, ChangeListener {
         add(txtGenero);
         // combo box Carrera
         List<String> listaCar = new ArrayList<>();
-        datos("carrera").stream().forEach((x) -> listaCar.add(x.split(";")[1]) );
+        Utilitario util = new Utilitario();
+        util.datos("carrera").stream().forEach((x) -> listaCar.add(x.split(";")[1]) );
         carreras = new JComboBox(listaCar.toArray());
         carreras.setBounds(160,250,150,20);
         add(carreras);
@@ -108,7 +109,7 @@ class PanelMenu extends JPanel implements ActionListener, ChangeListener {
         add(txtGenero);
         // combo box Semestre
         List<String> listaSem = new ArrayList<>();
-        datos("semestre").stream().forEach((x) -> listaSem.add(x.split(";")[1]) );
+        util.datos("semestre").stream().forEach((x) -> listaSem.add(x.split(";")[1]) );
         semestres = new JComboBox(listaSem.toArray());
         semestres.setBounds(160,290,150,20);
         add(semestres);
@@ -128,27 +129,7 @@ class PanelMenu extends JPanel implements ActionListener, ChangeListener {
     }
 
 
-    public List<String> datos(String flat){
 
-        List<String> carreras = new ArrayList<>();
-        FileReader FR;
-        try {
-            if(flat.equals("semestre")){
-                 FR=new FileReader("./txt/datosSemestre.txt");
-            }else{
-                 FR = new FileReader("./txt/datosCarrera.txt");
-            }
-            BufferedReader BR = new BufferedReader(FR);
-            String dato = "";
-            while ((dato = BR.readLine() ) != null ){
-                carreras.add(dato);
-            }
-
-        }catch (IOException e){
-            JOptionPane.showMessageDialog(null,"error al leer semestres " + e.getLocalizedMessage());
-        }
-        return carreras;
-    }
 
     public void stateChanged(ChangeEvent e) {
